edenapi_server: rename EdenApiContext to EdenApiServerContext
Summary: Later in this stack, we're going to have to introduce a few more context types, including Mononoke's per-session `CoreContext` as well as an LFS-server inspired per-request `RequestContext`. To make the scope of the `EdenApiContext` more clear (namely, that unlike the other contexts, this one persists for the entirety of the server's execution), let's rename this to `EdenApiServerContext`. This mirrors the naming convention for these contexts used in the LFS server.

Reviewed By: xavierd

Differential Revision: D19863296

fbshipit-source-id: 8ad785070328523d0beaf824c86c7350ff6a2697