fix thrift timeouts in `eden restart`
Summary:
Fix the `check_health()` function to always set a timeout when querying for
EdenFS's health.  Originally we used to always set a default timeout of 60
seconds when creating thrift connections to EdenFS, but this was removed in
D5942205.  In practice we ideally really want a handful of specific thrift
calls (e.g., 'checkOutRevision()`, `getScmStatusV2()`) to have extremely high
timeouts, but most other calls should have fairly short timeouts.

For now this ensures that we apply a 3 second timeout by default when checking
for EdenFS health.  The `edenfsctl status` call did explicitly set a 15 second
timeout, but other commands like `edenfsctl clone` and `edenfsctl restart`
would also check for health and were not applying their own timeout.

Also add thrift timeout for the `initiateShutdown()` call when doing a full
restart in `edenfsctl restart`

Reviewed By: chadaustin

Differential Revision: D20130405

fbshipit-source-id: c59118dbcafc2ed0d29206e33891f1a58da8c05f