add a TelemetryLogger interface
Summary:
Add a `TelemetryLogger` class that abstracts the mechanism we use to log
telemetry samples.  This makes it possible to plug in alternative
implementations.

This includes 3 initial implementations of this class:
* `ExternalTelemetryLogger` logs samples by calling an external command
* `LocalTelemetryLogger` logs JSON samples to a local file
* `NullTelemetryLogger` simply discards all samples

This also moves some of the helper code for constructing telemetry samples
from the `EdenInstance` class and into `TelemetryLogger`.

Reviewed By: genevievehelsel

Differential Revision: D20112725

fbshipit-source-id: dbe24952a92fe548631fc169f146cc14008a7bb6